Fix updateNode being called with undefined rootNode
When redux `dispatch()` is called before `getState()` has ever been called, `updateNode()` (in the custom `creatStore()` function) will be called with `rootNode` still undefined.

This wasn't catched in tests as they coincidentally always called `getState()` (for the cache setup) before dispatching. The added test here reproduces the bug that I was seeing when that's not the case.